Refactor, remove yarn
This is an effort to make this repo look remarkably similar to
github.com/actions/cache in structure, to facilitate housing a variety
of different actions in a single repository.

It also removes yarn - I'd like to KISS and stay close to actions/cache
as a set of best practices.