Move setup action to root dir
In order to publish to GH Marketplace, we need to have an action.yml in
the root dir.

Let's take github.com/actions/cache as an example. The base "cache"
action is in the root dir, and the additional "save" and "restore"
actions are under /save and /restore respectively.

We'll copy that here - massdriver-cloud sets up the cli, then you can
use massdriver-cloud/image_push etc.